backend/pyproject.toml
[project]
name = "mantenimiento-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.*",
  "uvicorn[standard]==0.30.*",
  "sqlmodel==0.0.22",
  "psycopg[binary]==3.2.*",
  "alembic==1.13.*",
  "python-multipart",
  "python-jose[cryptography]==3.3.*",
  "passlib[argon2]==1.7.*",
  "redis==5.0.*",
  "pydantic-settings==2.4.*",
  "slowapi==0.1.*",
  "pyotp==2.9.*"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []

[tool.black]
line-length = 100


